# public/.well-known/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Base OS Agent API",
    "version": "1.0.0",
    "description": "Read-only HTTP API for Base OS — onchain score, portfolio, radar, and analytics on Base (chain 8453). See /.well-known/SKILL.md for agent usage patterns.",
    "contact": {
      "url": "https://app-base-os.vercel.app"
    }
  },
  "servers": [
    {
      "url": "https://app-base-os.vercel.app",
      "description": "Production"
    }
  ],
  "paths": {
...
    "/api/transaction-trays": {
      "get": {
        "operationId": "getTransactionTrays",
        "summary": "Agent transaction tray definitions for tip and onchain games",
        "parameters": [
          {
            "name": "game",
            "in": "query",
            "schema": { "type": "string", "enum": ["grid646", "battleship"] }
          },
          {
            "name": "room",
            "in": "query",
            "schema": { "type": "string", "pattern": "^\\d+$" }
          }
        ],
        "responses": {
          "200": { "description": "Tray bundle with calldata actions" }
        }
      }
    },
...
  },
...
}